src/dj_alpha/requirements.txt
spotipy
python-dotenv
librosa
matplotlib
